[GRO-25] Fix incorporation date and entity type data gathering (#14)
## Problem
We are not gathering correctly the entity type and business age when people skip the business registration search. And even when they do not skip it, we need to ensure data is consistent across columns.

## Solution
We decided that we are going to ask for required fields if they skip the business registry search step. Please look at the evidence videos for this.

## Changes
* There was a database trigger that was using different labels for business age that were inconsistent with the ones we must ask. That was dropped
* Changes in form as seen in the videos
* Changes in the backend to ensure that we pass the fields to the database as needed.

## Evidence
* Skipping business registry search: https://www.loom.com/share/9271e2f5525943e586520d4abfce9951
* Not skipping business registry search: https://www.loom.com/share/aa122ea4988441429bb3cad77c814d42