[GRO-18] Add upload bank statements path (#13)
Problem

One problem regarding RFI is that some customers can't or are not willing to use the Flinks connection. For them, we need to enable a path form them to manually upload the bank statements

Changes

There's a new bank_statements table. Since application id is not set immediately, we use a UUID to link the bank statements together, and we use that as a folder for the uploads
UI changes to make this possible. See evidence for that
Added api endpoint to upload files
Added a migration for creating the bucket (that's the most declarative way to handle bucket creation in Supabase). Bucket is private so you cannot download publicly for now. We need to figure out best way to expose this to our internal users.
Evidence

https://www.loom.com/share/ccfe52d6f8d942b885e8cda4587bd75d?sid=d9e4db87-cc21-4a3f-8ee4-90b27def3670