[GRO-51] Allow early flow termination when user is not authorized to secure financing (#20)
## Problem

We have an RFI requirement from Journey to ensure that customers are actually eligible
to secure financing. We assumed that they required to be owners or own 50%> of the corporation,
but that's up for debate depending on the corporate structure. So now, if they meet the requirement we simply ask them if they own more than 50% but the answer is not a blocker.

Also, if they do not meet the requirements, we should simply terminate the application.

## Changes

* Simply change the text of step 3 to reflect the new requirements and terminate the application if they say now. 
* Instead of big buttons we added a simpler yes/no selection because if they say "yes" we show them the new question in the same step (to avoid adding more steps)

## Evidence
https://www.loom.com/share/3c698daf0cd34bd9b12311a47016d381?sid=819d890a-b747-4479-b82a-fd33a86cad8e